back/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["VictorGabrielMO <victorgabrielengsis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.17"
faster-whisper = "^1.0.3"
langchain = "0.0.259"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
